Added Tasks and Fields, altered a few file names, here SkillRepository is now using jpa to map skills by field, things are looking up!

+++ src/main/java/io/github/raphael_vi/skillkeeper/repositories/FieldsRepository.java
@@ -2,6 +2,6 @@ package io.github.raphael_vi.skillkeeper.repositories;
 import io.github.raphael_vi.skillkeeper.entities.Field;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FieldsRepository extends CrudRepository<Field, Long> {
+public interface FieldsRepository extends CrudRepository<Field, Integer> {
 
 }
